Moves NLC functions from dbManager to nlcconfig
