+++ test/seedDb.test.js
@@ -13,6 +13,7 @@ const initDb = require('../src/lib/initDb');
 const testDb = require('./setupTestDb');
 
 const nlcFile = 'test/resources/mock.seed.json';
+const nlcConfig = require('../src/lib/nlcconfig');
 
 // Passing arrow functions to mocha is discouraged: https://mochajs.org/#arrow-functions
 // return promises from mocha tests rather than calling done() - http://tobyho.com/2015/12/16/mocha-with-promises/
@@ -49,30 +50,30 @@ describe('Testing initial load of database', function() {
 
 		it('should test database class responses with seed data from a JSON file format', function() {
 			assert(this.db, 'db should be initialized');
-			return this.db.getClassEmitTarget('app.list').then((tgt) => {
+			return nlcConfig.getClassEmitTarget('app.list').then((tgt) => {
 				expect(tgt.target).to.eql('app.list');
-				return this.db.getClassEmitTarget('app.start');
+				return nlcConfig.getClassEmitTarget('app.start');
 			}).then((tgt) => {
 				expect(tgt.target).to.eql('app.start.js');
-				return this.db.getClassEmitTarget('app.test.1');
+				return nlcConfig.getClassEmitTarget('app.test.1');
 			}).then((tgt) => {
 				expect(tgt.target).to.eql('test.1.js');
-				return this.db.getClassEmitTarget('app.test.2');
+				return nlcConfig.getClassEmitTarget('app.test.2');
 			}).then((tgt) => {
 				expect(tgt.target).to.eql('test.2.js');
-				return this.db.getClassEmitTarget('app.test.3');
+				return nlcConfig.getClassEmitTarget('app.test.3');
 			}).then((tgt) => {
 				expect(tgt.target).to.eql('test.3.js');
-				return this.db.getClassEmitTarget('app.test.4');
+				return nlcConfig.getClassEmitTarget('app.test.4');
 			}).then((tgt) => {
 				expect(tgt.target).to.eql('test.4.js');
-				return this.db.getClassEmitTarget('app.test.5');
+				return nlcConfig.getClassEmitTarget('app.test.5');
 			}).then((tgt) => {
 				expect(tgt.target).to.eql('test.5.js');
-				return this.db.getClassEmitTarget('app.test.6');
+				return nlcConfig.getClassEmitTarget('app.test.6');
 			}).then((tgt) => {
 				expect(tgt.target).to.eql('test.6.js');
-				return this.db.getClassEmitTarget('app.test.7');
+				return nlcConfig.getClassEmitTarget('app.test.7');
 			}).then((tgt) => {
 				expect(tgt.target).to.eql('test.7.js');
 			});
@@ -82,11 +83,11 @@ describe('Testing initial load of database', function() {
 		it('should test database parameter responses with seed data from a JSON file format', function() {
 			assert(this.db, 'db should be initialized');
 			// app.test.2 uses a reference parameter.value
-			return this.db.getClassEmitTarget('app.test.2').then((tgt) => {
+			return nlcConfig.getClassEmitTarget('app.test.2').then((tgt) => {
 				expect(tgt.parameters.length).to.eql(2);
 				expect(tgt.parameters[0].values).to.exist;
 				expect(tgt.parameters[0].values.length).to.eql(5);
-				return this.db.getClassEmitTarget('app.test.5');
+				return nlcConfig.getClassEmitTarget('app.test.5');
 			}).then((tgt) => {
 				expect(tgt.parameters.length).to.eql(1);
 				expect(tgt.parameters[0].values).to.exist;
